log LP table at epoc x, log h/m at epoc x+1, get hit rate of LP for epoc x+1

+++ core/memory_subsystem/parametric_dram_directory_msi/memory_manager.h
@@ -129,6 +129,6 @@ namespace ParametricDramDirectoryMSI
          std::shared_ptr<cache_helper::CacheHelper> cacheHelper;
          std::shared_ptr<Helper::PCStatHelper> PCStatCollector;
          Helper::Counter epocCounter;
-         int debugEpoc;
+         UInt64 debugEpoc;
    };
 }
